* Refactored AbstractMotionPath out of MotionProfile to separate path solving from profile solving. * Simplified "PnP class" path solver heuristics completed.
